requirements.txt
google-api-python-client
textblob
Flask
Flask-Cors
python-dotenv
matplotlib
pandas
wordcloud